Made Lessons Functional, Added SQL Files
Filled out the functions in the LessonsController.

Added all the Lesson views.

Added an .SQL folder in the root directory that contains 2 folders.
The first contains a file that can be used to create the entire DB from a single file.
The second contains a file for each table of the DB so one can selectively choose which tables they want to import.

Fixed various routes and redirects using the url() function.