Added Bootstrap MultiSelect, Modified Readme
Added many-to-many relationship between Courses and Modules.

Added required js and css files to use the Bootstrap Multiselect tool.

Modified the README to include an introduction to Phanon and instructions to set up and development enviroment.

Modified Course views to allow for creating links between Courses and Modules.

Fixed some routes on the dashboard to use the url() function.