Object Relationships, CodeMirror Functions
Changed the data type of the code columns in the exercises table to be text instead of string.

Added all relationships between the course, module, lesson, and exercise objects.

Changed all the views to reflect this relationship.

On each objects index page, you can now see counts of what it contains and counts of what it is contained in.

On each objects show page, there are now links to any objects it contains.
There is also a list of any objects that contains it.

Made each objects show page follow the same format for consistency.

Added javascript functions to make making codeMirror stuff easier.

The function makeClassCodeMirror(sel) function allows you to pass in a css class as a string. Any textarea elements on the page with that class will be converted into a CodeMirror.

The function makeRunButton(btn_id) allows you to pass in an id of a button element and it will add an onclick event to it to call run.

The function run() was changed to find the first instance of a CodeMirror textarea on a page to pull the code from to transpile instead of being hard-coded.

Fixed the login and register routes on the dashboard.