Course Full View, Bug Fixes
Added course full view start. This view will be used to view everything in a course at once, on a single page. This will include its modules, lessons, projects, and exercises.

Added route for couse fullview

Fixed a bug in the CoursesController where the a variable name was changed in the view and not in the store and update functions.

Fixed module relation in the Lesson model

Added details and the bottom of every objects show view