Controller Fixes, Updated DB Seeder
Modified the if statement to check if an array exists and if it has something in it to just use the empty() function.

Modified the Store and Update functions of the LessonsController, ModulesController, and CoursesController to properly save and update the containing objects.

Fixed a copy-paste error on the exercises table migration that was giving the table a 'module_id' column instead of the correct 'lesson_id' column.

The DB Seeder now creates some sample objects to play around with. This should save time in testing so these objects dont have to be manually created.

Added code to the Course fullview to display everything contained in the course. This will need to be modified for a cleaner UI.

Fixed the index and show pages for Exercises to correctly show which lesson it is in.