Exercise Ordering within Lessons, DBSeeder Addition, Course Open/Close Dates, Function Renaming
Renamed existing "clone" functions to "copy" because clone is an existing function.

Added a createCourse function to the DBSeeder to create a course with relationships between the objects.

Renamed "deep_copy" functions to "deepCopy" to abide by PSR-1 convention.

Added open_date and close_date fields to the course object migration.

Gave names to existing routes.

Changed all instances of "/index" routes to be just "/".

Removed the createClone routes because they were overriding the create routes. They will be added again once a fix is found.

Added jquery-ui files and included references to them in the layout.

Added a nullable foreign key field to exercises for its previous_exercise_id. This will be used to store the order of exercises within a lesson.

Modified the create and edit view of the lessons to have a drag-and-droppable list element for a user to set the order of exercises within a lesson.

Added a orderedExercises function to the Lesson model that returns an array of containing exercises in their specified order.

Added a nextExercise function to the Lesson model that takes in an exercise id and returns the exercise that comes directly after it.