Concepts Functionality
Concepts are now fully functional.

Made it so the modules relation function in the Concept model will be called unorderedModules.

Added a function called modules to the Concept model that will return its modules in the correct order.

Added the nextModule function and removeModule function to the Concept model.

Modified the removeConcept function in the Course model to save the result of $this->concepts() to a variable.
This was to fix an error where end($this->concepts()) was throwing an error because references can't be passed into the end() function.

Renamed ConceptController to ConceptsController to match the naming of the other controllers.

Filled the resource functions of the ConceptsController as well as the corresponding views.
It was pretty much just copy/pasted from the CoursesController and views.

In the CoursesController, removed the check for concepts in the store and update function.
This was done because there could be a case where a course is created without concepts already in it.

Reworded some comments in the CoursesController to be more descriptive.