Lesson/Project Retrival for Modules, View Fixes, Additions to DBSeeder
Made it so the lessons relation function in the Module class will be called unorderedLessons.

Made it so the projects relation function in the Module class will be called unorderedProjects.

Added a function called lessons to the Module class that will return its lessons in the correct order.

Added a function called lessonsAndProjects to the Module class that will return an array of all its lessons and projects.
The lessons and projects will be in the correct order they need to be completed in.

Added nextLesson function to the Module class.

Changed the show function in the ModulesController to make use of the lessonsAndProjects function.

Added a concept and five filled-out modules to the DBSeeder file.

Removed the Open Date label from the lessons show view since it was removed from the Lesson class.

Updated the modules index view with the new names for the lessons and projects relations.

Modified the modules show view to show the details for the containing concept instead of the containing course.

Removed the Close Date label from the modules show view.

Modified the modules show view from showing the lessons and projects in seperate sections, to showing all lessons
and projects in the same section and in their correct order.