Addition of Enums, DBSeeder Additions, PSR-2 Formatting
Added base Enum class.

Added Permissions enum class.
Includes an admin permissions and CRUD permissions for courses, concepts, modules, lessons, exercises, and projects.

Added Roles enum class.
Includes admin, power user, teacher, teaching assistant, student, and observer roles and descriptions for each one.

Added the 'clearance' middleware to the ConceptsController, LessonsController, ExercisesController, and ProjectsController.

Modified the formatting of the PagesController to follow PSR-2 guidelines.
Added function comments to the functions of the PagesController.
Added use statements for the Roles and Permissions enums to the PagesController.
Changed the admin permission check in the PagesController to use the enum.

Modified the ClearanceMiddleware to use Permissions enum.

Modified the formatting of the DatabaseSeeder to follow PSR-2 guidelines.
The seeder will now create permissions based on the entries of the Permissions enum instead of manually create each one.
The seeder will now create roles based on the entries of the Roles enum instead of manuallu creating each one.
Added the creation of two more teacher users and teaching assistant users.

Changed the permissions used in the @can directives in the dashboard view.

Modified the permissions index view to show each permission in a unorderedList element instead of inline.