Added Ability to Add Users to Courses
Removed the course full view; its view, controller function, and route.
Removed the link to the course full view from the course show view.

Added a query to get all users in the create function of the CourseController.

Create a migration for the pivot table course_user.
This will be for the many-to-many relation between courses and users.
It contains the course_id, user_id, and the role_id that user has in the course.

Added users() relationship function to the Course model that will get all users in the course.
Added students() relationship function to the Course model that will get all student users in the course.
Added assistants() relationship function to the Course model that will get all teaching assistant users in the course.
Added teachers() relationship function to the Course model that will get all teacher users in the course.

Added addUsersAsRole() function to the Course model that will take in an array of user ids and a role id.
It will add every user to the course as the passed-in role.

Added inCourses() relationship function the User model that will get all courses the user is in.
Added takingCourses() relationship function to the User model that will get all courses the user is a student in.
Added assistingCourses() relationship function to the User model that will get all courses the user is a teaching assistant in.
Added teachingCourses() relationship function to the User model that will get all courses the user is a teacher in.

Added the creation of a few more student users to the DBSeeder.
Added the creation of a obersver user to the DBSeeder.

Filled out the users of the course created in the createTestCourse function of the DBSeeder.
Added a teacher, a teaching assistant, and three students.
Added the Student role to the two teaching assistant users being created in the DBSeeder.

Added sections to display a courses teacher, teaching assistants, and students to the course show view.

Added multiselect controls for adding students, teaching assistants, and teachers to the course.

Added some test code to the dashboard function of the PagesController that shows how to get different categories of courses for a user.