New Function for MultiSelect, Updated CoursesController for Editing
Changed the addUsersAsRole function of the Course class to use the syncWithoutDetaching function instead of attach.

Modified the edit and update functions of the CoursesController to allow for editing the students, tas, and teachers in a course.

Wrote a function called makeMultiSelect in the create-edit-form.js file that turns a select element into a JQuery-UI multiselect.
It takes in the id of a select element and the default list display text.

Changed all instances of creating a multiselect within views to use the new function.
The affected files are: concepts\create, concepts\edit, courses\create, courses\edit, lessons\clone, lessons\create, lessons\edit, modules\clone, modules\create, and modules\edit.