More Generalized Functions, Updated Course View, Added Tables to Save Progress
-User Model-
Added ->withPivot('role_id') to the courses relation.
Added getCoursesByRole function that will return an array of courses the user has the specified role in.

-Course Model-
Added ->withPivot('role_id') to the users relation.
Added getUsersByRole function that will return an array of users with the specified role.

-Course Show View-
Removed the teachers, students, and teaching assistants sections from the Course show view.
Added a table to the Course show view that will display all users in the course and what role they have.

-Pages Controller-
Removed some extraneous code.

-Dashboard View-
Added the courses close date to the table.
Added a "Role" column to see what role the logged in user has for the course.
Moved the @endcan for the course.delete permission down to encompass the Delete button created as a form so that users without that permission cannot see the delete button.

-Migrations-
Added table migrations for users_exercise_completion and users_project_completion.