Course Create/Edit Changes, Permission Checking, Site Cleaning
Added permission checking for the Create button on each objects index view page.

Removed any instance of the Clone button on each objects show page.
The cloning functions need to be rewrote and tested again.

Fixed a bug where the Edit and Delete button on a lessons show page were not checking if the logged in user owned that item.

Fixed the project show view to correctly parse any html that could be written into the project's prompt.
Changed some html elements to be more semantically correct.

Added addUserAsRole function to the Course model that will add a single user to the course as the specified role.

Modified course create and edit views to use dropdowns to select role instead of having elements to add users of pre-determined roles.
Modified the create and edit functions of the CoursesController to create a roles array and pass it to their respective views.
Modified the store and update functions of the CoursesController to make use of new form input.

Added updateTable, makeRoleDropdown, and addUserRoleInputToForm to the create-edit-form.js file.
These functions are used to populate the table with selected users, create the dropdown for roles, and add the input the form when its submitted.

Removed the students, assistants, and teachers relationship functions from the Course model.

Made it so when a Course is edited, it redirects to the Course show page instead of the Course index page so you can review the changes that were made.