Changed Return of python.py, Refactored codewindow.js
-python.py
Changed how the messages for the tests are returned to javascript.
There is now a boolean for whether the test was successful.
If a test was successful, there is also now be a message returned.

Changed how __out_string is initialized. No longer contains "None" as the first element.

The function test_out will no longer only overwrite the first value of __out_string.
It will contain an element for each test_out call.

There will now be a message added to problems for each call to test_in, instead of a single message for all of them.

Fixed an issue in the function __UNIVERSAL_VALIDATOR that was causing test_type to not return a message correctly.

-codewindow.js
Changed the wording of the open multi-line comment error message.

Removed the printError function.

Refactored the run function to be less complicated.
Functions were added to accomplish this.

Added a makeTestMessagesArray function that will take in the return from python.py and parse it into a workable array of objects.
Each object will have a passed boolean property and a message string property.
The message property will contain the result from each test and passed will indicate whether the test was successful or not.

Added printTestMessages function that will take in the testMessages array and output the result from each test.

Added printPythonErrors function that takes in the return variable from Skulpt and displays the Python error message.

Added various comments throughout to describe better how everything is working.