Updated Properties on Each Object, Code Cleaning
Renamed user_id column of Courses, Concepts, Modules, Lessons, Projects, and Exercises to owner_id.
Renamed the user relationship function in each model to be owner.
Updated all instances of user to owner in all the respective controllers and views.
Updated the property desciptions of the six objects to be more consistent and include owner_id.
Updated the databaseSeeder file to use the new name.
Updated the reader.blade.php file to use the new name.

Added an updated_by column to each of the six objects.
Added the property desciptions of the six objects to include updated_by.

Added a solution column to Exercises.
Added the property description for solution to the Exercise model.

Removed extraneous functions from the User model.

Renamed the inCourses function of the User model to enrolledCourses.

Updated the LucidChart ERD to be up-to-date with the current state of the database.