ExerciseProgress and ProjectProgress Model and Controllers, Tile Coloring On ExerciseDash, Bug Fixes
Added "exercise.autocomplete" to the Permissions enum.
Updated the databaseSeeder to add the "exercise.autocomplete" permission to the Teacher role.

Renamed users_exercise_completion to exercise_progress.
Renamed users_project_completion to project_progress.

Modified the migrations for users_exercise_progress and users_project_progress tables to include updated columns.

Added a model for ExerciseProgress.
Added property descriptions for all properties.
Added a completed function that will return whether the exercise is completed or not.

Added a model for ProjectProgress.
Added property descriptions for all properties.

Added some test code to the exercise function of the codeController that passes the users in the course to the views.
An array of exercise completions is also being created and passed to the views.
This is an early prototype concept to colorize the exerciseDash based on whether the logged-in user has completed exercises or not.

Added a completed and current css class to the site.css for coloring of the tiles on the ExerciseDash.
Added an if statement to the exerciseDash.blade.php that will give a tile the class completed if an exercise is completed.

Modified the codearea/module.blade.php to pass the module_completion array to the exerciseDash and the users array to the codeEditorWithTests.
Added code to the codeEditorWithTests.blade.php that will create a form to autocomplete exercises if the logged-in user has the permission for it.

Added a route for the complete function of the ExerciseProgressController.

Fixed a bug on the courses function of the user model where the query used to find owned courses was using the wrong column name on the courses table.
Fixed a bug on the dashboard function of the PagesController where the User models courses function wasn't being used correctly.

Removed some commented out old code from the dbSeeder.