Save ExerciseProgress When on Code Run
-ExerciseProgress Migration
Modified the completion_date field to be nullable.

-ExerciseProgressController
Added logic to completed() that will update the completion_date field if the user has already completed the exercise.

Added a save() function that will handle the AJAX call that occurs when the Run button is pressed on a code editor page.

Added a saveProgress function that will handle saving the ExerciseProgress object to the database.
If the user has progress saved, the appropiate fields will get updated.
If not, one will be created and saved to the database.

-codewindow.js
Added logic to the run() function that will check if the exercise was completed correctly or not and save it to a boolean.
The run function will now save the users code when it its ran without Python errors.

-codearea/codeEditorWithTests.blade.php
Added a script that defines a save(completed) function.
This function takes in a boolean for whether the exercise was completed correctly or not and makes an AJAX call to ExerciseProgressController@save.

-Routes
Added the route for ExerciseProgressController@save.