Modified TestMessages Array, Bug Fixes on Coding Views
Added a check to see if the $user variable exists in the @can("exercise.autocomplete") section of the codeEditorWithTests view.

Added a variable to store the li class for the tiles on the exercise dash.
It will only get filled if the module completion variable is passed to the view.

Added a getTests() function to the codewindow.js file that gets the Python test code and tests to run and returns them as a single string.
It can take in a boolean to indicate to include the Python test code or not.
This allows to pass false and just get the tests to run back.

Renamed the attribute on the testMessages array from "passed" to "success".
Added an attribute to the testMessages array called "test" that will contain the test that was called as a string.