Improved Exercise Saving, Better Code Flow
Moved the saveProgress function from ExerciseProgressController to the model.
Significantly improved the logic of saveProgress.
It will now correctly fill every field correctly based on whether the exercise was completed correctly or not.

Removed testing strings from the save function in the ExerciseProgressController.

Moved the definition of the javascript save function that makes the AJAX call from a script tag on the codeEditorWithTests view to the codewindow.js file.
Added global variables to the codeEditorWithTests view that is filled with PHP variables and allows the javascript save function to work.