Coloring Current Exercise Tile
Changed the current class in the site.css to modify the border color of an exercise class element instead of the background color.
Removed changing the color of an <a> element in a current class element.
Modified the hover of an exercise class element to no longer change the border color.
Now, there is a clear indication of what the current exercise is and which ones are completed or not.

Added logic to the run function in codewindow.js to add the completed class if the current exercise was finished correctly.

Added logic on the exerciseDash to add the current class to the current exercises tile.

Added code to codearea/module to pass the current exercise object to the exerciseDash.