Addition of Utility Functions to Models, View and Component Rewrite, Code Editor Javascript Rewrite, New Routes
Added a completed function to the Concept model that returns whether the concept is completed or not for the logged-in user.

Added a getProgressForUser function to the Exercise model that gets the corresponding ExerciseProgress object for the logged-in user.

Added a getProgress funciton to the ExerciseProgress model that will fetch the ExerciseProgress object for a given exercise and user.

Removed some old code from the CodeController.
Added a newexercise function to the CodeController that is called from the new route "/newexercise/{exercise_id}".

Added a newsave function to the ExerciseProgressController that is called from an AJAX call to save an exercise to the database.

Added a completed function to the Lesson model that returns whether the lesson is completed or not for the logged-in user.

Added a completed function to the Module model that returns whether the module is completed or not for the logged-in user.
Added a completion function to the Module model that creates an array of the contained lessons and exercises and whether the logged-in user has completed them yet.
This is used on the exerciseDash to color the tiles accordingly.

Added completeExerciseForUser and completeLessonForUser functions to the DatabaseSeeder.
They can be used to automatically create progress when the db is seeded.

Removed the call to js/codewindow.js from the layouts/app view.
This file is only used on pages with a code editor so that is the only time it will be loaded.

Completly rewrote the views and components for code editor pages. It is now more modular and significantly more readable.
The old views and such are all still intact until it is agreed upon to switch to these new files.

Rewrote the codewindow.js file to be more modular and clean. The old one still exists so the rewrite is named codeeditor.js.

Added a scriptbundle for the codewindow.js and codeeditor.js files.
The scriptbundle is used in the appropiate spots on code editors and the sandbox.

Added routes for the new exercise view pages and saving.
The new route for doing exercises only takes in an exercise id.
The old one took in exercise id and lesson id which was pointless.
The lesson id can be obtained from the exercise and you will never have a case where you try to look at an exercise id and lesson id where the exercise isnt in that lesson.

Removed the Questions.txt file from the _Documents folder.