Started Replacing Old Views With New Ones, Route Updates
Replaced the old exercise and project functions in the CodeController with updated versions.

Added nextIncompleteExercise function to the Lesson model.
It will return either the first incomplete exercise in a lesson or the first exercise in the lesson.

Added fullIDE class to projectEditor partial view.
Added project start code to the view.

Changed the url for lessons on the flow page to go to the updated pages.
Changed the url for projects on the flow page to go to the updated pages.

Removed any unused or unnecassary routes.
Gave names to the import routes.
Added new routes for the updated exercise and project pages.

Fixed formatting of ImportController.