Saving Projects, Loading Users Code for Exercises and Projects, Added Reset Button
Deleted files from views/codearea that have been replaced.
Review.blade.php was kept for the filtering code.

Renamed codearea2 folder to codearea.

Deleted extra codewindow.js file in the resources/assets folder.

Deleted extra save function from ExerciseProgressController.

Deleted old routes.
Fixed it so there are two routes for both code/exercise and code/project.
One goes to the page where the item can be attempted and the other saves the users attempt to the database.

Created the ProjectProgressController.
Added save function that is used for the AJAX route to save a Project.

Added latestContents function to the ExerciseProgressController that will return the users code from when they saved most recently.

Added passing of the users saved code to the views when code/exercise or code/project is visited.

Added saveProgress function to the ProjectProgressController that saves a users attempt at a project to the database.
A new ProjectProgress will be created every time the user runs the code.

Added makeResetButton function to codeeditor.js that takes in a button id and makes it reset the code to the original start code when pressed.
Added saveProjectCode function that makes the AJAX call to save the project progress to the database.
Rewrote some logic in runCode for when to call saveExerciseCode and saveProjectCode.

Added logic to get a users latest code for a project or exercise and pass it through the views.
When the user loads up a project or exercise, if they have tried it before, their code will be loaded in the editor.

Added a Reset button to codearea/buttons partial.
The exercise or projects start_code is contained in the data attribute.