Start of Adding Partners, New has_partners Field on Projects, New id function in Enums/Roles
Added id function to Enums/Roles to get the id of a roles by passing in its name.

Removed the unique constraint on the naming of projects.

Projects now have a has_partners boolean field that determines whether the project will be worked on solo or as a team.
Added hasPartners function the Project model that returns whether the project has partners enabled or not. Passing true into the function will return the boolean as a string if desired.
Modified ProjectController and views to allow for editing the has_partners field of projects.

Added partners function to ProjectsController for the route /project/{id}/partners.
It is used to view all teams that exist in the course the project is in and what teams have been assigned to this project.

Added Team model that groups users together.
Defined relationship between Teams and the projects they are assigned to.
Defined relationship between Teams and the members in the team.
Defined relationship between Teams and the course the team belongs to.

Created projects/partners.blade.php view for viewing the teams assigned to a project.

Added TeamProjectProgress model that will contain a teams work on a project.
Added TeamsController.php.
Added TeamsProjectProgressController.php.

Created database migrations to make the teams, team_users, project_teams, and team_project_progress tables.