Team Permission Creation, New Users in DBSeed, Team Creation Functions, Project Assignment Functions
Added CRUD permissions for teams to the Permissions enum.

Added teams function to CourseController that goes to a page where all teams in a course can be seen.
If the user has permission to create new teams, a form will be there to allow that.

Added validation to partners function of the ProjectsController that checks whether the user is in the course the project belongs to.

Added createTeam function to TeamsController that takes in a list of students and creates a team consisting of them.

Added assignTeams function to Project model that takes in an array of team_ids and assigns them to the project using the sync function.
Sync means any existing teams will be erased and only the teams in the array will be assigned anymore.

Added addMember, addMembers, makeTeam, and assignProject functions to the Team model.
addMember adds a single member to a team
addMembers takes in an array of users and calls addMember for each of them
makeTeam takes in a course_id and an array of users and creates the team
assignProject takes in a project_id and assigns the team to work on it

Added isAdmin function to user model that returns true if the user is an admin.

Modified the DBSeeder to give team permissions to certain roles.
Added more student users to be created.
Added students to the course being created.

Added courses/teams.blade.php view that lists any teams that belong to the course.
A user with the correct permission can create a new team from this view as well.

Modified the projects/partners.blade.php view so the Project name is a link to do the project.
This view now only shows the teams that are currently assigned to the project.
A user with the correct role within the course can press the Assign Random Teams button.
This button shuffles all Student users within the course into teams and assigns the project to them.

Added partials/teamsTable.blade.php view that takes in an array of teams and displays them in a tabular format.

Added new lines to any files that didnt end in them.