Logic for Unique Team Creation and Validation
Added logic to the createTeam and assignRandomTeams functions of the TeamsController that checks if the team trying to be created already exists. Only unique teams are allowed within courses.

Added checkIfTeamExistsInCourse function to Team model. It will take in a course_id and an array of user_ids and check if the team already exists in the course. If it does, it will return it, else it returns false.

Moved the form on courses/teams.blade.php view to above the teamsTable.