New Course Create Page, Course Flow Bug Fixes
Modified LoginController to clear the session on logout.

Created FlowController for the flow.authoring pages.
These control the page where a user can create a course using the flow layout.
The create function calls the page to create a new course.
The store function is called when the flow course create form is submitted.
Added routes for the create and store functions of the FlowController.

Created resources/views/flow/authoring folder to contain any views related to course authoring.
Created flow/authoring/create.blade.php file that contains the main view for the flow course create page.
Created flow/authoring/store.blade.php file that contains a testing view that shows the course structure after the course create form is submitted.

Changed the path of the Course Create button on the dashboard to the new course create page.

Created js/forms/edit_course.js file to handle the logic of the dynamic flow course create page.
It has functions that can add concept, module, lesson, and project elements as well as being able to delete them.
Created a scriptbundle for the js file for course create form.

Added styling of list items of an element with the class .components.authoring within the courseFlow element.

Fixed any bugs on the flow/lesson.blade.php view caused by objects that were assumed to be there.
Fixed any bugs on the flow/module.blade.php view caused by objects that were assumed to be there.
Fixed a bug in the CompletionStats function of the Module model that would occur if the query returned nothing.