Improved Course Flow Load Time
Added an alias for the Permissions Enum so it can be used on blade views without needing manually load the php file.

Modified the unorderedModules relationship function of the Concept model to be ordered by the previous_module_id.
Modified the unorderedConcepts relationship function of the Course model to be ordered by the previous_concept_id.
Modified the unorderedExercises relationship function of the Lesson model to be ordered by the previous_exercise_id.
Modified the unorderedLessons relationship function of the Module model to be ordered by the previous_lesson_id.
Modified the unorderedProjects relationship function of the Module model to be ordered by the previous_lesson_id.

Added the show function to the FlowController to handle showing the flow view of a course.
It contains code to eager load only the required fields of the course for a performance boost compared to the old flow.
Moved the flow function from the PagesController to the FlowController.

Added the components function to the Module model that returns a collection of lessons and projects in the correct order with only certain columns selected.
This is used on the new flow view for eager loading.
Added testLessons and testProjects relationship functions to the Module model for use with the components function.

Modified the ExerciseCount function of the Module model to return 0 if the query returned null.

Commented out the .project:before and .lesson:before parts of the site.css for now.
Added some styling to the end of site.css for the a tag of the components list and the name span of the components list.

Created js/pages/course_flow.js file to handle any js that needs to happen on the course_flow page.
Added toggleExpandCollapse function that takes in a module_id and handles the logic of animating it collapsing and expanding.

Added a scriptbundle for the course_flow.js file.

Created new view partials for the course flow modeled directly after the old ones.
This was just to breakdown the old ones in an attempt to understand why loading was taking so long.
Also, to clear out any unused, commented out, or old code.