Project Saving for Teams, Links to Course and Project Team Pages
Added projects function to the course object that returns an array of all the projects in the course in the order they appear in.

The course teams page now shows all projects within the course with links to the page to edit what teams are assigned to each project.

The save function of the ProjectProgressController now checks if a team is logged in and if it is, saves the project for the team instead of an individual user.

Added comments to each function of the TeamsController.
Changed the message shown when a team is created within a course to sound better.
Changed the name of the showLoginForm function to loginForm.

Added code to explicitly state that there aren't timestamps for the TeamProjectProgress object.
Added the saveProgress function to the TeamProjectProgress object that saves a project for a team.

Added the toggleEditable function to the js/forms/edit_course.js file that can toggle whether an element is editable or not.

Adjusted some spacing on the codearea/projectEditor.blade.php file.

Modified the course flow authoring page to resemble the course flow page more.
Modified some elements to use the toggleEditable js function.

Added a link to the course teams page from the course flow page when the logged-in user has the correct permission.

Removed unneeded code from the pages/dashboard.blade.php file.