Edit Functionality on the Course Flow Page
Made it so the Create Course button on the dashboard creates a new course and redirects to the courses flow page.

Added create functions for concepts, modules, lessons, and projects to the FlowController.
These will set default values for things like names and dates while retaining the order within its containing object.

Changed which courses will show on the dashboard.

Added some logic to prevent the CurrentExercise function within the module model from breaking.

Added a button to the flow/course.blade.php file that toggles Edit Mode on the flow page.
When Edit Mode is enabled, any create buttons will be visible.

Added some javascript to the flow/course.blade.php file.
The functions that were added are:
 toggleVisibilityByClass() - Toggles whether an element is visible or not if it has the class name passed into the function.
 toggleEditButtonText() - Changes the button text of the edit mode button from "Enable Edit Mode" to "Turn Off Edit Mode".
 createConcept() - Sends an AJAX call to the server to create a concept and return its partial.
 createModule() - Sends an AJAX call to the server to create a module and return its partial.
 createLesson() - Sends an AJAX call to the server to create a lesson and return its partial.
 createProject() - Sends an AJAX call to the server to create a project and return its partial.

Added a button to the flow/index.blade.php file to create a new concept through an AJAX call.
Added a button to the flow/concept.blade.php file to create a new module through an AJAX call.
Added two buttons to the flow/module.blade.php file to create a new project or lesson through an AJAX call.

Added routes for the AJAX calls to create the objects on the course flow page.

Minor formatting on flow partial views.