Added Middleware to Controllers Requiring A User, Better JS on the Course Flow Page
Added 'auth' and 'clearance' middleware to CodeController to require be logged in for all functions except sandbox.
Removed unused newexercise function.

Added 'auth' and 'clearance' middleware to CodeController to require be logged in for all functions except index and about.

Commented out a call to a success() function in the js/actions.js file.

Removed the hard coded class from the toggleVisibilityByClass function of the js/pages/course_flow.js file.
Added expandModules function that finds all collapsed modules on the course flow page and expands them.
Added a call to expandModules when Edit Mode is turned on.